Add some CRF support
This patch adds the basic bits of CRF support. The avtp_crf_pdu struct
is defined in its own header, as it's specific to this protocol.
All the functions required to set and get the values in general struct
are implemented.

The actual timestamps will be part of the payload of this struct and
it's not yet clear what the best way to handle them would be, so like
for the AAF case, they are left for applications for the moment.